+++ lib/djv/ModelsPy/FilesModel.cpp
@@ -64,6 +64,7 @@ namespace djv
                     const std::shared_ptr<FilesModelItem>&>(&FilesModel::add))
                 .def("add", py::overload_cast<
                     const std::vector<std::shared_ptr<FilesModelItem> >&>(&FilesModel::add))
+                .def("move", &FilesModel::move, py::arg("fromIndex"), py::arg("toIndex"))
                 .def("close", py::overload_cast<>(&FilesModel::close))
                 .def("close", py::overload_cast<int>(&FilesModel::close), py::arg("index"))
                 .def("closeAll", &FilesModel::closeAll)